requirements.txt
blosc2==2.0.0
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.3
msgpack==1.0.4
numexpr==2.8.4
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
py-cpuinfo==9.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tables==3.8.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
typing_extensions==4.4.0
urllib3==1.26.14
